+++ packages/main/test/pages/base/AccessibilityTextsHelper.html
@@ -4,7 +4,7 @@
 <head>
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>AriaLabelHelper</title>
+    <title>AccessibilityTextsHelper</title>
     <script src="../%VITE_BUNDLE_PATH%" type="module"></script>
 </head>
